+++ haul-fe/src/components/MobileLayout.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import theme from "../assets/theme/Theme";
+import theme from "../styles/theme/Theme";
 
 const Background = styled.div`
   width: 100%;
